[src] changes to handle large corpus
Moved everything to classes to better facilitate in storing lots of
data. The intent was to handle large corpus. Also wanted to handle
stashing the TEXT elements for generation. Python crashed too many
times and I eventually took the TEXT storing out for now.

The basic idea though was to be able to process a very large corpus,
though the results indicate the corpus size has little to do with the
quality of the results.